Revert to a compilable and functional build
Up until this point, this project was mainly a local playground. The latest WIP state was not finished and therefore not compilable. I've now decided to create a repo for it and maintain upgrades/changes via Gitflow. This means that the latest doodles are now on a different branch, and this commit is meant to revert back to a previously working version